+++ src/Root/Data/Root.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * @copyright  2018 Jack Clayton
+ * @license    MIT
+ */
+
+namespace Jstewmc\Gravity\Root\Data;
+
+use SplFileInfo;
+
+class Root extends SplFileInfo
+{
+    // nothing yet
+}
