+++ src/Root/Exception/Exception.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * @copyright  2018 Jack Clayton
+ * @license    MIT
+ */
+
+namespace Jstewmc\Gravity\Root\Exception;
+
+use Jstewmc\Gravity\Exception as GravityException;
+
+abstract class Exception extends GravityException
+{
+    // nothing yet
+}
