+++ src/Root/Exception/NotFound.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * @copyright  2018 Jack Clayton
+ * @license    MIT
+ */
+
+namespace Jstewmc\Gravity\Root\Exception;
+
+class NotFound extends Exception
+{
+    public function __construct()
+    {
+        $this->message = "Root directory not found";
+    }
+}
