+++ tests/Root/Service/FindTest.php
@@ -1,21 +1,15 @@
 <?php
 /**
- * The file for the find-filesystem service tests
- *
- * @author     Jack Clayton <clayjs0@gmail.com>
  * @copyright  2018 Jack Clayton
  * @license    MIT
  */
 
-namespace Jstewmc\Gravity\Filesystem\Service;
+namespace Jstewmc\Gravity\Root\Service;
 
 use PHPUnit\Framework\TestCase;
 
 /**
- * Tests for the find-filesystem service
- *
- * @since  0.1.0
- * @todo
+ * @group  root
  */
 class FindTest extends TestCase
 {
@@ -31,6 +25,6 @@ class FindTest extends TestCase
 
     // public function testInvokeReturnsProjectDirectoryIfComposer(): void
     // {
-    //     // how the heck do I test this?
+    //     // hmm, how the heck do I test this?!
     // }
 }
